src/test/kotlin/year2024/Day01Test.kt
package year2024

import Utils.Companion.cases
import Utils.Companion.readTestFile
import org.assertj.core.api.Assertions.*
import org.junit.jupiter.params.*
import org.junit.jupiter.params.provider.*

class Day01Test {

    companion object {
        @JvmStatic
        fun data() = cases(
            arrayOf(
                Arguments.of("year2024/Day01_sample", 11, 31),
            ),
            Arguments.of("year2024/Day01_personal", 2264607L, 19457120L)
        )
    }

    @ParameterizedTest(name = "File {0}, part 1 answer = {1}, part 2 answer = {2}")
    @MethodSource("data")
    fun test(file: String, part1Answer: Long, part2Answer: Long) {
        val contents = readTestFile(file)
        assertThat(Day01().part1(contents)).isEqualTo(part1Answer)
        assertThat(Day01().part2(contents)).isEqualTo(part2Answer)
    }
}
